chore: Move ConversationCannotDecryptSystemMessageCellDescription & ConversationStartedSystemMessageCellDescription into its own file
