+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/ConversationSystemMessageCell.swift
@@ -194,42 +194,6 @@ private extension ConversationLike {
 
 // MARK: - Descriptions
 
-final class ConversationStartedSystemMessageCellDescription: NSObject, ConversationMessageCellDescription {
-
-    typealias View = ConversationStartedSystemMessageCell
-    let configuration: View.Configuration
-
-    var message: ZMConversationMessage?
-    weak var delegate: ConversationMessageCellDelegate?
-    weak var actionController: ConversationMessageActionController?
-
-    var showEphemeralTimer: Bool = false
-    var topMargin: Float = 0
-
-    let isFullWidth: Bool = true
-    let supportsActions: Bool = false
-    let containsHighlightableContent: Bool = false
-
-    let accessibilityIdentifier: String? = nil
-    var conversationObserverToken: Any?
-
-    init(message: ZMConversationMessage, data: ZMSystemMessageData) {
-        let color = LabelColors.textDefault
-        let iconColor = IconColors.backgroundDefault
-        let model = ParticipantsCellViewModel(font: .mediumFont, largeFont: .largeSemiboldFont, textColor: color, iconColor: iconColor, message: message)
-
-        actionController = nil
-        configuration =  View.Configuration(title: model.attributedHeading(),
-                                            message: model.attributedTitle() ?? NSAttributedString(string: ""),
-                                            selectedUsers: model.selectedUsers,
-                                            icon: model.image())
-        super.init()
-
-        accessibilityLabel = configuration.message.string
-    }
-
-}
-
 class ConversationIgnoredDeviceSystemMessageCellDescription: ConversationMessageCellDescription {
 
     typealias View = ConversationNewDeviceSystemMessageCell
@@ -268,144 +232,6 @@ class ConversationIgnoredDeviceSystemMessageCellDescription: ConversationMessage
         }
 
         return .markdown(from: string, style: .systemMessage)
-    }
-
-}
-
-class ConversationCannotDecryptSystemMessageCellDescription: ConversationMessageCellDescription {
-    typealias View = ConversationCannotDecryptSystemMessageCell
-    let configuration: View.Configuration
-
-    static fileprivate let resetSessionURL: URL = URL(string: "action://reset-session")!
-
-    var message: ZMConversationMessage?
-    weak var delegate: ConversationMessageCellDelegate?
-    weak var actionController: ConversationMessageActionController?
-
-    var showEphemeralTimer: Bool = false
-    var topMargin: Float = 0
-
-    let isFullWidth: Bool = true
-    let supportsActions: Bool = false
-    let containsHighlightableContent: Bool = false
-
-    let accessibilityIdentifier: String? = nil
-    let accessibilityLabel: String?
-
-    init(message: ZMConversationMessage, data: ZMSystemMessageData, sender: UserType) {
-        let icon: UIImage
-        if data.systemMessageType == .decryptionFailedResolved {
-            icon = StyleKitIcon.checkmark.makeImage(size: 16, color: IconColors.foregroundCheckMarkInSystemMessage)
-        } else {
-            icon = StyleKitIcon.exclamationMark.makeImage(size: 16, color: IconColors.foregroundExclamationMarkInSystemMessage)
-        }
-
-        let title = ConversationCannotDecryptSystemMessageCellDescription
-            .makeAttributedString(
-                systemMessage: data,
-                sender: sender
-            )
-
-        configuration = View.Configuration(icon: icon,
-                                           attributedText: title,
-                                           showLine: false)
-        accessibilityLabel = title.string
-        actionController = nil
-    }
-
-    func isConfigurationEqual(with other: Any) -> Bool {
-        guard let otherDescription = other as? ConversationCannotDecryptSystemMessageCellDescription else {
-            return false
-        }
-
-        return configuration.attributedText == otherDescription.configuration.attributedText
-    }
-
-    // MARK: - Localization
-
-    private static let BaseLocalizationString = "content.system.cannot_decrypt"
-    private static let IdentityString = ".identity"
-
-    private static func makeAttributedString(systemMessage: ZMSystemMessageData, sender: UserType) -> NSAttributedString {
-
-        let messageString = self.messageString(systemMessage.systemMessageType, sender: sender)
-        let resetSessionString = self.resetSessionString()
-        let errorDetailsString = self.errorDetailsString(
-            errorCode: systemMessage.decryptionErrorCode?.intValue ?? 0,
-            clientIdentifier: (systemMessage.senderClientID ?? "N/A"))
-
-        var components: [NSAttributedString]
-
-        switch systemMessage.systemMessageType {
-        case .decryptionFailed:
-            components = [
-                messageString
-            ]
-
-            if systemMessage.isDecryptionErrorRecoverable {
-                components.append(resetSessionString)
-            }
-        case .decryptionFailedResolved:
-            components = [
-                messageString,
-                errorDetailsString
-            ]
-        case .decryptionFailed_RemoteIdentityChanged:
-            components = [
-                messageString
-            ]
-        default:
-            fatal("Incorrect cell configuration")
-        }
-
-        return components.joined(separator: NSAttributedString(string: "\n"))
-    }
-
-    private static func localizationKey(_ systemMessageType: ZMSystemMessageType) -> String {
-        let localizationKey: String
-        switch systemMessageType {
-        case .decryptionFailed:
-            localizationKey = BaseLocalizationString
-        case .decryptionFailedResolved:
-            localizationKey = BaseLocalizationString + "_resolved"
-        case .decryptionFailed_RemoteIdentityChanged:
-            localizationKey = BaseLocalizationString + "_identity_changed"
-        default:
-            fatal("Incorrect cell configuration")
-        }
-
-        return localizationKey
-    }
-
-    private static func messageString(_ systemMessageType: ZMSystemMessageType, sender: UserType) -> NSAttributedString {
-
-        let name = sender.name ?? ""
-        var localizationKey = self.localizationKey(systemMessageType)
-
-        if sender.isSelfUser {
-            localizationKey += ".self"
-        } else {
-            localizationKey += ".other"
-        }
-
-        return NSMutableAttributedString.markdown(from: localizationKey.localized(args: name), style: .systemMessage)
-    }
-
-    private static func resetSessionString() -> NSAttributedString {
-        let string = L10n.Localizable.Content.System.CannotDecrypt.resetSession
-
-        return NSAttributedString(string: string.localizedUppercase,
-                                  attributes: [.link: resetSessionURL,
-                                               .foregroundColor: UIColor.accent(),
-                                               .font: UIFont.mediumSemiboldFont])
-    }
-
-    private static func errorDetailsString(errorCode: Int, clientIdentifier: String) -> NSAttributedString {
-        let string = L10n.Localizable.Content.System.CannotDecrypt.errorDetails(errorCode, clientIdentifier)
-
-        return NSAttributedString(string: string.localizedUppercase,
-                                  attributes: [.foregroundColor: LabelColors.textDefault,
-                                               .font: UIFont.mediumFont])
     }
 
 }
